Move _nexus to third party directory
