+++ pyproject.toml
@@ -19,7 +19,7 @@ classifiers = [
     "Operating System :: OS Independent",
 ]
 
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 dependencies = [
     "numpy",
     "scipy",
@@ -67,7 +67,6 @@ fallback_version = "0.0.0.dev"
 line-length = 88
 indent-width = 4
 exclude = [
-    "_nexus",
     "_ext",
 ]
 
@@ -94,9 +93,8 @@ parallel = true
 
 [tool.coverage.report]
 omit = ['*_test.py',
-    'weblogo/seq_io/_nexus/*.py',
     'weblogo/utils/deoptparse.py',
-    'weblogo/_ext/*.py']
+    'weblogo/_ext/*']
 
 exclude_lines = [
   '\#\s*pragma: no cover',
@@ -114,7 +112,6 @@ exclude_lines = [
 
 [tool.mypy]
 files = 'weblogo'
-exclude = ["weblogo/seq_io/_nexus/", "weblogo/_ext"]
 
 # Suppresses error about unresolved imports
 ignore_missing_imports = true
