+++ weblogo/_ext/__init__.py
@@ -1,3 +1,3 @@
 #
-# Imported code
+# Third party imported code
 #
