+++ weblogo/htdocs/manual.html
@@ -488,7 +488,7 @@
                             <dt>
                                 <h4>Dependencies</h4></dt>
                             <dd>
-                                WebLogo version 3 is written in python. It is necessary to have <a href="http://www.python.org/download/">Python 3.8 or later</a> and the extension package
+                                WebLogo version 3 is written in python. It is necessary to have <a href="http://www.python.org/download/">Python 3.11 or later</a> and the extension package
                                 <a href="http://www.scipy.org/Download">numpy</a> installed before WebLogo will run. WebLogo also requires a recent version of <a href="http://www.cs.wisc.edu/~ghost/">ghostscript</a> to create PNG and PDF output, and <a href="http://www.cityinthesky.co.uk/opensource/pdf2svg">pdf2svg</a> to generate SVG output.
                             </dd>
                             <dt>
